Adjusted publish tests and workflow
