+++ tests/publish_image/conftest.py
@@ -1,0 +1,10 @@
+from os import getenv
+
+import pytest
+from click.testing import CliRunner
+
+
+@pytest.fixture(scope="package")
+def cli_runner() -> CliRunner:
+    runner = CliRunner()
+    return runner
