+++ tests/publish_image/test_cli_and_env.py
@@ -1,10 +1,8 @@
-import pytest
 from click.testing import CliRunner, Result
 
 from build.publish import main
 
 
-@pytest.mark.usefixtures("cleanup_images")
 def test_missing_options_and_env(cli_runner: CliRunner):
     result: Result = cli_runner.invoke(main)
     assert result.exit_code == 2
